Get poverty and inequality statistics

Get poverty and inequality statistics

get_stats(
  country = "all",
  year = "all",
  povline = 1.9,
  popshare = NULL,
  fill_gaps = FALSE,
  group_by = NULL,
  welfare_type = c("all", "income", "consumption"),
  reporting_level = c("all", "national", "urban", "rural"),
  version = NULL,
  api_version = "v1",
  format = c("json", "csv", "rds"),
  simplify = TRUE,
  server = NULL
)

Arguments

country

character: A vector with one or more country ISO 3 codes or 'all'

year

integer: A vector with one or more years or 'all'

povline

numeric: Poverty line

popshare

numeric: Proportion of the population living below the poverty line

fill_gaps

logical: If TRUE, will interpolate / extrapolate values for missing years

group_by

character: If used result will be aggregated for predefined sub-groups

welfare_type

character: Welfare type

reporting_level

character: Geographical reporting level

version

character: Data version. See get_versions()

api_version

character: API version

format

character: Response format

simplify

logical: If TRUE (default) the response is converted to a data frame

server

character: Server. For WB internal use only

Value

tibble or list

Examples

if (FALSE) {
# One country-year
res <- get_stats(country = "AGO", year = 2000)

# All years for a specific country
res <- get_stats(country = "AGO", year = "all")

# All countries and years
res <- get_stats(country = "all", year = "all")

# All countries and years w/ alternativ poverty line
res <- get_stats(country = "all", year = "all", povline = 3.2)

# Fill gaps for years without available survey data
res <- get_stats(country = "all", year = "all", fill_gaps = TRUE)

# Proportion living below the poverty line
res <- get_stats(country = "all", year = "all", popshare = .4)
}